chore: Props 및 페이지네이션 타입 분리
