+++ src/types/EditPagination.ts
@@ -1,0 +1,4 @@
+export type ArrowButton = {
+  $right?: string;
+  $left?: string;
+};
